+++ src/zuckermaas.h
@@ -7,21 +7,21 @@
 
 class ZuckerMaas : public BossOfBoris {
 private:
-    std::vector<float> theta; //policy parameters
-    std::vector<float> zt; //current momentum
-    std::vector<float> delta; //current step
-    const float alpha; //learning rate 0.2 a good value?
-    const float beta; //momemntum importance factor
+    std::vector<double> theta; //policy parameters
+    std::vector<double> zt; //current momentum
+    std::vector<double> delta; //current step
+    const double alpha; //learning rate 0.2 a good value?
+    const double beta; //momemntum importance factor
     int t;
 public:
-    ZuckerMaas(float alpha);
+    ZuckerMaas(double alpha);
     virtual BorisGoal getGoal(const State &currentState);
-    //void setParameterVector(const std::vector<float> vector){ parameters = vector; }
+    //void setParameterVector(const std::vector<double> vector){ parameters = vector; }
 private:
-    float evaluateAction(BorisGoal action);
-    std::vector<float> getFeatures(const State& currentState, const BorisGoal& action);
+    double evaluateAction(BorisGoal action);
+    std::vector<double> getFeatures(const State& currentState, const BorisGoal& action);
     void initializeParameterVector(int size);
-    float calculateQuality(const std::vector<float> &parameterVector, const std::vector<float> &features);
+    double calculateQuality(const std::vector<double> &parameterVector, const std::vector<double> &features);
 };
 
 #endif // ZUCKERMAAS_H
